requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2.9
django-crispy-forms==1.13.0
django-redis==5.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
httpie==2.6.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
redis==4.0.1
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.6
sqlparse==0.4.2
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
wrapt==1.13.3